requirements.txt
amqp==2.2.1
billiard==3.5.0.3
boto==2.48.0
boto3==1.4.7
botocore==1.7.7
celery==4.1.0
dj-database-url==0.4.2
Django==1.11.5
django-celery-beat==1.0.1
django-celery-results==1.0.1
djangorestframework==3.6.4
docutils==0.14
gunicorn==19.7.1
jmespath==0.9.3
kombu==4.1.0
olefile==0.44
Pillow==4.2.1
python-dateutil==2.6.1
pytz==2017.2
redis==2.10.6
s3transfer==0.1.11
six==1.10.0
vine==1.1.4
